requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==4.1.0
Django==4.0.3
django-ckeditor==6.2.0
django-filebrowser==4.0.1
django-grappelli==3.0.3
django-js-asset==2.0.0
django-quill-editor==0.1.40
django-summernote==0.8.20.0
django-tinymce==3.4.0
jsmin==3.0.1
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
